Reorganize the Game 2 includes and namespaces
