+++ Prototyping/Games/2 - Circle shoots Cubes/CircleCubeSpecHelper.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "GameSpecHelper.h"
+
+namespace Prototyping::Games::CircleCube {}
+
+using namespace Prototyping::Games::CircleCube;
