+++ Prototyping/Games/2 - Circle shoots Cubes/include/Prototyping/Games/2DCircleCube/Core/GameCircleCubeInputHandler.Spec.cpp
@@ -2,7 +2,7 @@
 
 #include "GameCircleCubeInputHandler.h"
 
-#include "GameSpecHelper.h"
+#include "CircleCubeSpecHelper.h"
 
 Describe("GameCircleCubeInputHandler") {
     it("moves the circle to a left-clicked cell", []() {
