+++ Prototyping/Games/2 - Circle shoots Cubes/include/Prototyping/Games/2DCircleCube/Core/GameCircleCubeInputHandler.h
@@ -2,7 +2,7 @@
 
 #include "GameCircleCube.h"
 
-namespace Prototyping {
+namespace Prototyping::Games::CircleCube {
 
     class GameCircleCubeInputHandler {
         GameCircleCube& _game;
