+++ Prototyping/Games/2 - Circle shoots Cubes/include/Prototyping/Games/2DCircleCube/Qt/QtGameCircleCubeScene.Spec.cpp
@@ -6,7 +6,7 @@
 #include <QGraphicsSceneMouseEvent>
 #include <QGraphicsView>
 
-#include "GameSpecHelper.h"
+#include "CircleCubeSpecHelper.h"
 
 using namespace Prototyping;
