+++ Prototyping/Games/2 - Circle shoots Cubes/include/Prototyping/Qt/ParticleExplosion.h
@@ -1,0 +1,3 @@
+#pragma once
+
+namespace Prototyping {}
